moved items from array to object for better organization and performance

+++ src/stores/items.ts
@@ -8,20 +8,32 @@ export interface IItem {
     quantity: number,
     planned: boolean,
     completed: boolean | null
-    shoppinglist_id?: string
+    shoppinglist_id?: number
     id?: string
 }
 
+type ItemStore = {
+  [index: number]: IItem[]
+}
+
 export const useItemsStore = defineStore('items', () => {
-  const items = ref<IItem[]>([])
+  const items = ref<ItemStore>({})
   
   async function addItem(item: IItem, listId: string): Promise<void> {
 
     try {
       const request = await api.post(`/shoppinglist/${listId}/item`, item)
-
+      
       if(request.status == 201) {
-        items.value.push(request.data)
+        const newItem: IItem = request.data
+        if(!newItem.shoppinglist_id) return
+
+        if(items.value[newItem.shoppinglist_id]) {
+          items.value[newItem.shoppinglist_id].push(newItem)
+        } else {
+          items.value[newItem.shoppinglist_id] = [newItem]
+        } 
+
       } else {
           console.log("Something bad happened!")
         }
@@ -32,35 +44,57 @@ export const useItemsStore = defineStore('items', () => {
   }
 
   function addItems(newItems: IItem[]){
-    items.value.push(...newItems)
+
+    for(const item of newItems) {
+
+      if(item.shoppinglist_id) {
+
+        if(items.value[item.shoppinglist_id]) {
+
+          items.value[item.shoppinglist_id].push(item)
+        } else {
+
+          items.value[item.shoppinglist_id] = [item]
+        }
+
+      }
+    }
+
   }
 
-  function getItems(listId: string): Array<IItem> {
-    const listItems: IItem[] = items.value.filter(item => item.shoppinglist_id == listId)
-    return listItems
+  function getItems(listId: number | string): Array<IItem> {
+
+    if(typeof listId == "string") {
+      return items.value[parseInt(listId)] ?? []
+    }
+    return items.value[listId] ?? []
+
   }
 
-  async function loadItems(listId: string | undefined): Promise<void> {
-    if(items.value.length <= 0 && listId) {
+  async function loadItems(listId: number | undefined): Promise<void> {
+    if(!listId) return
+
+    if(!items.value[listId]) {
 
       console.log("Making a server request to get items")
 
       const request = await api.get(`/shoppinglist/${listId}/item`)
       
       if(request.status == 200) {
-        items.value = [...items.value, ...request.data]
+        items.value[listId] = request.data
       }
     }
   }
 
   async function removeItem(item: IItem | undefined) {
     if(!item) return
+    if(!item.shoppinglist_id) return
 
     try {
       const request  = await api.delete(`/shoppinglist/${item.shoppinglist_id}/item/${item.id}`)
       
       if(request.status == 200) {
-        items.value = items.value.filter(curItem => curItem.id != item.id)
+        items.value[item.shoppinglist_id] = items.value[item.shoppinglist_id].filter(curItem => curItem.id != item.id)
       }
 
     } catch (err: any) {
@@ -71,6 +105,7 @@ export const useItemsStore = defineStore('items', () => {
 
   async function completeItem(item: IItem): Promise<void> {
     if(!item) return
+    if(!item.shoppinglist_id) return
 
     try {
       const request = await api.put(
@@ -79,8 +114,8 @@ export const useItemsStore = defineStore('items', () => {
       )
 
       if(request.status == 200) {
-        const itemIndex = items.value.findIndex(i => i.id == item.id)
-        items.value[itemIndex].completed = !item.completed
+        const itemIndex = items.value[item.shoppinglist_id].findIndex(i => i.id == item.id)
+        items.value[item.shoppinglist_id][itemIndex].completed = !item.completed
       }
